requirements.txt
aiogram==2.25.1
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
asyncpg==0.27.0
attrs==22.2.0
Babel==2.9.1
certifi==2022.12.7
charset-normalizer==3.0.1
frozenlist==1.3.3
greenlet==2.0.2
idna==3.4
magic-filter==1.0.9
multidict==6.0.4
Pillow==9.4.0
python-dotenv==0.21.1
pytz==2022.7.1
SQLAlchemy==2.0.3
typing_extensions==4.5.0
yarl==1.8.2
